chore: fix file names in csv parser
