+++ packages/pieces/community/csv/src/index.ts
@@ -1,6 +1,6 @@
 import { PieceAuth, createPiece } from '@activepieces/pieces-framework';
-import { parseCSVTextAction } from './lib/actions/convert-json-to-csv';
-import { unparseCSVTextAction } from './lib/actions/convert-csv-to-json';
+import { unparseCSVTextAction } from './lib/actions/convert-json-to-csv';
+import { parseCSVTextAction } from './lib/actions/convert-csv-to-json';
 
 export const csv = createPiece({
   displayName: 'CSV',
