+++ packages/pieces/community/csv/src/lib/actions/convert-csv-to-json.ts
@@ -1,16 +1,15 @@
 import { createAction, Property } from '@activepieces/pieces-framework';
-import { UnparseConfig } from 'papaparse';
-import { unparseCSVObject } from '../utils';
+import { parseCSVFile } from '../utils';
 
-export const unparseCSVTextAction = createAction({
-  name: 'convert_json_to_csv',
-  displayName: 'Convert JSON to CSV',
+export const parseCSVTextAction = createAction({
+  name: 'convert_csv_to_json',
+  displayName: 'Convert CSV to JSON',
   description:
-    'This function reads a JSON file and converts it into a CSV file format.',
+    'This function reads a CSV string and converts it into JSON array format.',
   props: {
-    csv_object: Property.Json({
-      displayName: 'CSV JSON',
-      defaultValue: {},
+    csv_text: Property.LongText({
+      displayName: 'CSV Text',
+      defaultValue: '',
       required: true,
     }),
     has_headers: Property.Checkbox({
@@ -33,16 +32,12 @@ export const unparseCSVTextAction = createAction({
     }),
   },
   async run(context) {
-    const { csv_object, has_headers, delimiter_type } = context.propsValue;
-    const config: UnparseConfig = {
+    const { csv_text, has_headers, delimiter_type } = context.propsValue;
+    const config = {
       header: has_headers,
       delimiter: delimiter_type === 'auto' ? '' : delimiter_type,
       skipEmptyLines: true,
     };
-
-    const results = unparseCSVObject(csv_object, config);
-    console.debug('Unparse results', results);
-
-    return results;
+    return parseCSVFile(csv_text, config);
   },
 });
