--- packages/pieces/community/csv/src/lib/csv.mdx
@@ -1,14 +1,0 @@
----
-title: 'CSV'
-description: 'Convert from / to json'
----
-
-## Triggers
-
-TRIGGERS
-
----
-
-## Actions
-
-ACTIONS
